Rename new NFTs files
