+++ background/lib/simple-hash.ts
