+++ background/redux-slices/selectors/index.ts
@@ -4,5 +4,5 @@ export * from "./keyringsSelectors"
 export * from "./signingSelectors"
 export * from "./dappSelectors"
 export * from "./uiSelectors"
-export * from "./nftsSelectors_update"
+export * from "./nftsSelectors"
 export * from "./abilitiesSelectors"
